requirements.txt
discord.py
python-dotenv
yt-dlp
spotipy
pynacl
davey
groq